Add --guess-common-names (#51)
Fixes #46 
leaving as draft until specs has been finalized, i.e. how to handle
common names with default value of `None` and the extent of common names
to use for inferring.

I decided that the least intrusive way of implementing was to hijack
`annotate_optionals` [and `annotate_named_params`, so I just append the
guessed names to those lists. I at first considered implementing this in
`leave_Param`, and could move back there if that's preferred. It took me
a while to parse `leave_Param` and I started adding comments for myself,
and figured I might as well leave them.